requirements.txt
rdkit==2023.09.1
streamlit==1.31.0
Pillow==10.2.0